// apps/api/src/types/qbittorrent.types.ts
export interface QBittorrentTorrent {
  hash: string;
  name: string;
  size: number;
  progress: number; // 0-1
  dlspeed: number; // bytes/s
  upspeed: number; // bytes/s
  priority: number;
  num_seeds: number;
  num_complete: number;
  num_leechs: number;
  num_incomplete: number;
  ratio: number;
  eta: number; // seconds
  state: string;
  seq_dl: boolean;
  f_l_piece_prio: boolean;
  category: string;
  tags: string;
  super_seeding: boolean;
  force_start: boolean;
  save_path: string;
  added_on: number; // unix timestamp
  completion_on: number;
  tracker: string;
  dl_limit: number;
  up_limit: number;
  downloaded: number;
  uploaded: number;
  downloaded_session: number;
  uploaded_session: number;
  amount_left: number;
  save_path_id?: number;
  time_active: number;
  seeding_time: number;
}

export interface QBittorrentServerState {
  connection_status: string;
  dht_nodes: number;
  dl_info_data: number;
  dl_info_speed: number;
  dl_rate_limit: number;
  up_info_data: number;
  up_info_speed: number;
  up_rate_limit: number;
  queueing: boolean;
  use_alt_speed_limits: boolean;
  refresh_interval: number;
}

export interface QBittorrentPreferences {
  save_path: string;
  temp_path_enabled: boolean;
  temp_path: string;
  scan_dirs?: Record<string, number>;
  download_path_enabled: boolean;
  download_path: string;
  export_dir_enabled: boolean;
  export_dir: string;
  mail_notification_enabled: boolean;
  mail_notification_sender: string;
  mail_notification_email: string;
  mail_notification_smtp: string;
  mail_notification_ssl_enabled: boolean;
  mail_notification_auth_enabled: boolean;
  mail_notification_username: string;
  mail_notification_password: string;
  autorun_enabled: boolean;
  autorun_program: string;
  queueing_enabled: boolean;
  max_active_downloads: number;
  max_active_torrents: number;
  max_active_uploads: number;
  dont_count_slow_torrents: boolean;
  slow_torrent_dl_rate_threshold: number;
  slow_torrent_ul_rate_threshold: number;
  slow_torrent_inactive_timer: number;
  max_ratio_enabled: boolean;
  max_ratio: number;
  max_ratio_act: number;
  listen_port: number;
  upnp: boolean;
  random_port: boolean;
  dl_limit: number;
  up_limit: number;
  max_connec: number;
  max_connec_per_torrent: number;
  max_uploads: number;
  max_uploads_per_torrent: number;
  stop_tracker_timeout: number;
  enable_piece_extent_affinity: boolean;
  bittorrent_protocol: number;
  limit_utp_rate: boolean;
  limit_tcp_overhead: boolean;
  limit_lan_peers: boolean;
  alt_dl_limit: number;
  alt_up_limit: number;
  scheduler_enabled: boolean;
  schedule_from_hour: number;
  schedule_from_min: number;
  schedule_to_hour: number;
  schedule_to_min: number;
  scheduler_days: number;
  dht: boolean;
  pex: boolean;
  lsd: boolean;
  encryption: number;
  anonymous_mode: boolean;
  proxy_type: number;
  proxy_ip: string;
  proxy_port: number;
  proxy_peer_connections: boolean;
  proxy_auth_enabled: boolean;
  proxy_username: string;
  proxy_password: string;
  proxy_torrents_only: boolean;
  ip_filter_enabled: boolean;
  ip_filter_path: string;
  ip_filter_trackers: boolean;
  web_ui_domain_list: string;
  web_ui_address: string;
  web_ui_port: number;
  web_ui_upnp: boolean;
  web_ui_username: string;
  web_ui_password: string;
  web_ui_csrf_protection_enabled: boolean;
  web_ui_clickjacking_protection_enabled: boolean;
  web_ui_secure_cookie_enabled: boolean;
  web_ui_max_auth_fail_count: number;
  web_ui_ban_duration: number;
  web_ui_session_timeout: number;
  web_ui_host_header_validation_enabled: boolean;
  bypass_local_auth: boolean;
  bypass_auth_subnet_whitelist_enabled: boolean;
  bypass_auth_subnet_whitelist: string;
  alternative_webui_enabled: boolean;
  alternative_webui_path: string;
  use_https: boolean;
  ssl_key: string;
  ssl_cert: string;
  web_ui_https_key_path: string;
  web_ui_https_cert_path: string;
  dyndns_enabled: boolean;
  dyndns_service: number;
  dyndns_username: string;
  dyndns_password: string;
  dyndns_domain: string;
  rss_refresh_interval: number;
  rss_max_articles_per_feed: number;
  rss_processing_enabled: boolean;
  rss_auto_downloading_enabled: boolean;
  rss_download_repack_proper_episodes: boolean;
  rss_smart_episode_filters: string;
  add_trackers_enabled: boolean;
  add_trackers: string;
  web_ui_use_custom_http_headers_enabled: boolean;
  web_ui_custom_http_headers: string;
  max_seeding_time_enabled: boolean;
  max_seeding_time: number;
  announce_ip: string;
  announce_to_all_tiers: boolean;
  announce_to_all_trackers: boolean;
  async_io_threads: number;
  banned_IPs: string;
  checking_memory_use: number;
  current_interface_address: string;
  current_network_interface: string;
  disk_cache: number;
  disk_cache_ttl: number;
  embedded_tracker_port: number;
  enable_coalesce_read_write: boolean;
  enable_embedded_tracker: boolean;
  enable_multi_connections_from_same_ip: boolean;
  enable_os_cache: boolean;
  enable_upload_suggestions: boolean;
  file_pool_size: number;
  outgoing_ports_max: number;
  outgoing_ports_min: number;
  recheck_completed_torrents: boolean;
  resolve_peer_countries: boolean;
  save_resume_data_interval: number;
  send_buffer_low_watermark: number;
  send_buffer_watermark: number;
  send_buffer_watermark_factor: number;
  socket_backlog_size: number;
  upload_choking_algorithm: number;
  upload_slots_behavior: number;
  upnp_lease_duration: number;
  utp_tcp_mixed_mode: number;
}